ml_app/requirements.txt
torch
flask
pillow
mysql-connector
opencv-python-headless
torchvision
colorspacious
colormath
gdown
gradio